Move site/congregation looks up stuff to the lookup repository, service, and controller.
