--- SignInCheckIn/MinistryPlatform.Translation/Repositories/Interfaces/ISiteRepository.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-using MinistryPlatform.Translation.Models.DTO;
-using Newtonsoft.Json.Linq;
-
-namespace MinistryPlatform.Translation.Repositories.Interfaces
-{
-    public interface ISiteRepository
-    {
-        List<MpCongregationDto> GetAll();
-    }
-}
