+++ SignInCheckIn/MinistryPlatform.Translation/Repositories/LookupRepository.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using Crossroads.Web.Common.MinistryPlatform;
 using MinistryPlatform.Translation.Models.DTO;
 using MinistryPlatform.Translation.Repositories.Interfaces;
@@ -54,6 +55,27 @@ namespace MinistryPlatform.Translation.Repositories
 
             return _ministryPlatformRestRepository.UsingAuthenticationToken(apiUserToken)
                 .Search<MpCountryDto>("1=1", columns);
+        }
+
+
+        /// <summary>
+        /// Gets a list of the Congregations in MP
+        /// </summary>
+        /// <returns></returns>
+        public List<MpCongregationDto> GetCongregations()
+        {
+            var apiUserToken = _apiUserRepository.GetToken();
+
+            var congregationColumnList = new List<string>
+            {
+                "Congregation_ID",
+                "Congregation_Name"
+            };
+
+            var congregations = _ministryPlatformRestRepository.UsingAuthenticationToken(apiUserToken)
+                .Search<MpCongregationDto>($"Available_Online = 1 AND (End_Date IS NULL OR End_Date > '{DateTime.Now:yyyy-MM-dd}')", congregationColumnList);
+
+            return congregations;
         }
     }
 }
