+++ SignInCheckIn/SignInCheckIn/Services/Interfaces/ILookupService.cs
@@ -7,5 +7,6 @@ namespace SignInCheckIn.Services.Interfaces
     {
         List<StateDto> GetStates();
         List<CountryDto> GetCountries();
+        List<CongregationDto> GetCongregations();
     }
 }
