--- SignInCheckIn/SignInCheckIn/Services/Interfaces/ISiteService.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using SignInCheckIn.Models.DTO;
-
-namespace SignInCheckIn.Services.Interfaces
-{
-    public interface ISiteService
-    {
-        List<CongregationDto> GetAll();
-    }
-}
