+++ SignInCheckIn/SignInCheckIn/Services/LookupService.cs
@@ -27,5 +27,12 @@ namespace SignInCheckIn.Services
             var countries = _lookupRepository.GetCountries();
             return Mapper.Map<List<MpCountryDto>, List<CountryDto>>(countries);
         }
+
+        public List<CongregationDto> GetCongregations()
+        {
+            var congregationDtos = _lookupRepository.GetCongregations();
+
+            return Mapper.Map<List<MpCongregationDto>, List<CongregationDto>>(congregationDtos);
+        }
     }
 }
