--- SignInCheckIn/SignInCheckIn/Services/SiteService.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using AutoMapper;
-using MinistryPlatform.Translation.Models.DTO;
-using MinistryPlatform.Translation.Repositories.Interfaces;
-using SignInCheckIn.Models.DTO;
-using SignInCheckIn.Services.Interfaces;
-
-namespace SignInCheckIn.Services
-{
-    public class SiteService : ISiteService
-    {
-        private readonly ISiteRepository _siteRepository;
-
-        public SiteService(ISiteRepository siteRepository)
-        {
-            _siteRepository = siteRepository;
-        }
-
-        public List<CongregationDto> GetAll()
-        {
-            var congregationDtos = _siteRepository.GetAll();
-
-            return Mapper.Map<List<MpCongregationDto>, List<CongregationDto>>(congregationDtos);
-        }
-    }
-}
